Extract dts bundle plugin to its own module
Move `getDtsBundlePlugin` and its `collectInlinedLibraries` /
`buildDtsPaths` helpers from `rollupConfig.mjs` into a sibling
`dtsBundlePlugin.mjs`. `rollupConfig.mjs` now just imports the
plugin factory; the dts-specific deps (`dts-bundle-generator`,
`typescript`) only load on the consumer side.
